continue setup greeting views
