+++ QuitSmoke/App/SceneDelegate.swift
@@ -12,9 +12,21 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     var window: UIWindow?
 
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-        // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
-        // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
-        // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
+        
+        let mainStoryboard: UIStoryboard = UIStoryboard(name: "Main", bundle: nil)
+        let defaults = UserDefaults.standard
+        if let _ = defaults.value(forKey: "UserData") {
+            if let navigationVC = mainStoryboard.instantiateViewController(withIdentifier: "navController") as? UINavigationController {
+                window?.rootViewController = navigationVC
+                window?.makeKeyAndVisible()
+            }
+        } else {
+            if let greetingVC = mainStoryboard.instantiateViewController(withIdentifier: "greetingVC") as? UIViewController {
+                window?.rootViewController = greetingVC
+                window?.makeKeyAndVisible()
+            }
+        }
+        
         guard let _ = (scene as? UIWindowScene) else { return }
     }
